requirements.txt
azure-functions
azure-identity
azure-storage-blob
azure-keyvault-secrets
requests
